sxgui2 interfaces

+++ source/gui/CSSstyle.h
@@ -187,10 +187,10 @@ namespace gui
 			float4_t getColor()const
 			{
 				return(float4_t(
-					(float)((m_value.i & 0xFF000000) >> 24) / 255.0f,
-					(float)((m_value.i & 0x00FF0000) >> 16) / 255.0f,
-					(float)((m_value.i & 0x0000FF00) >> 8) / 255.0f,
-					(float)((m_value.i & 0x000000FF)) / 255.0f
+					(float)(((UINT)m_value.i & 0xFF000000) >> 24) / 255.0f,
+					(float)(((UINT)m_value.i & 0x00FF0000) >> 16) / 255.0f,
+					(float)(((UINT)m_value.i & 0x0000FF00) >> 8) / 255.0f,
+					(float)(((UINT)m_value.i & 0x000000FF)) / 255.0f
 					));
 			}
